refactor weather components  to typescript
